OpStack now is serializable via Protobuf.
But, the implementation is hard to debug: right now, the objects' states are computed using a stored descriptor, but it should be the other way around.

Part of the reason for this is that JS/TS only allow one constructor per class, if the overloads have the same number of parameters... and some classes have only one state element to initialize.

Maybe I can get around this using the SerializableConstructor inferface?